Move PaymentProviderChannel, Configuration to montonio package and rename
