--- src/main/java/ee/tuleva/onboarding/payment/provider/PaymentProviderChannel.java
@@ -1,10 +1,0 @@
-package ee.tuleva.onboarding.payment.provider;
-
-import lombok.Data;
-
-@Data
-public class PaymentProviderChannel {
-  String accessKey;
-  String secretKey;
-  String bic;
-}
