+++ src/main/java/ee/tuleva/onboarding/payment/provider/montonio/MontonioPaymentChannel.java
@@ -1,0 +1,10 @@
+package ee.tuleva.onboarding.payment.provider.montonio;
+
+import lombok.Data;
+
+@Data
+public class MontonioPaymentChannel {
+  String accessKey;
+  String secretKey;
+  String bic;
+}
